feat(Load status): move status select component to Cargo item
